Split oscar.test.newfactories in multiple modules
Create a factory module per app and move them under the new
oscar.test.factories module.
