+++ tests/integration/offer/availability_tests.py
@@ -2,7 +2,7 @@ from django.test import TestCase
 
 from oscar.apps.offer import models
 from oscar.test.factories import create_order
-from oscar.test.newfactories import OrderDiscountFactory, UserFactory
+from oscar.test.factories import OrderDiscountFactory, UserFactory
 
 
 class TestAPerUserConditionalOffer(TestCase):
