+++ tests/unit/customer/alert_tests.py
@@ -3,7 +3,7 @@ from django.test import TestCase
 from oscar.apps.customer.models import ProductAlert
 from oscar.core.compat import get_user_model
 from oscar.test.factories import create_product
-from oscar.test.newfactories import UserFactory
+from oscar.test.factories import UserFactory
 
 
 User = get_user_model()
